+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Reject empty values when setting a secret
+
 ## [0.7.1] - 2026-02-08
 
 ### Changed